requirements.txt
Flask
pandas
numpy
torch
torchvision
Pillow
requests
MarkupSafe
gunicorn
